pyproject.toml
[project]
name = "forgent"
version = "0.4.0"
description = "Planning + knowledge layer for AI coding agents. Routes any task to the best of 60+ curated specialists, returns a structured PlanCard (steps, gotchas, success criteria) with progressive memory, and ships an optional Claude Code status line."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ali Al Aayedi", email = "alialaayedi@users.noreply.github.com" }]
maintainers = [{ name = "Ali Al Aayedi", email = "alialaayedi@users.noreply.github.com" }]
keywords = [
    "agents",
    "agent-orchestration",
    "ai-agents",
    "anthropic",
    "claude",
    "claude-code",
    "crewai",
    "langgraph",
    "llm",
    "llm-agents",
    "mcp",
    "mcp-server",
    "multi-agent",
    "orchestration",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.urls]
Homepage      = "https://github.com/alialaayedi/forgent"
Repository    = "https://github.com/alialaayedi/forgent"
Documentation = "https://github.com/alialaayedi/forgent/blob/main/docs/INTEGRATION.md"
Issues        = "https://github.com/alialaayedi/forgent/issues"
Changelog     = "https://github.com/alialaayedi/forgent/blob/main/CHANGELOG.md"
"Brand guide" = "https://github.com/alialaayedi/forgent/blob/main/docs/brand.md"
"MCP server"  = "https://github.com/alialaayedi/forgent/blob/main/docs/INTEGRATION.md"

[project.optional-dependencies]
langgraph = ["langgraph>=0.2.0", "langchain>=0.3.0", "langchain-anthropic>=0.2.0"]
crewai = ["crewai>=0.80.0"]
mcp = ["mcp>=1.0.0"]
all = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "crewai>=0.80.0",
    "mcp>=1.0.0",
]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
forgent = "forgent.cli:app"
forgent-mcp = "forgent.mcp_server:main"
forgent-statusline = "forgent.statusline:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"forgent.registry" = ["*.yaml"]
"forgent.registry.agents.claude_code" = ["*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["sources", ".venv", "build", "dist"]
